# back/planilla-rubik-back/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Planilla Rubik API",
    "version": "v1.0.0",
    "description": "API para el sistema Planilla Rubik, encargado de la gestión de proyectos, colaboradores, tecnologias, vacantes, configuraciones, notificaciones y usuarios.",
    "contact": {
      "name": "Equipo de Desarrollo Planilla Rubik"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3900/api",
      "description": "Servidor de Desarrollo Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "accessTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-access-token",
        "description": "Token de acceso (x-access-token) para rutas protegidas."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "fail",
              "error"
            ],
            "example": "fail"
          },
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo del error."
          },
          "details": {
            "type": "string",
            "description": "Detalles adicionales del error (opcional).",
            "example": "El campo 'email' es requerido."
          }
        },
        "required": [
          "status",
          "message"
        ]
      },
...
      "RoleEnum": {
        "type": "string",
        "enum": [
          "user",
          "admin",
          "moderator"
        ],
        "description": "Rol del usuario en el sistema."
      },
      "UserBase": {
        "type": "object",
        "description": "Información base de un usuario, sin incluir contraseña.",
        "properties": {
          "_id": {
            "type": "string",
            "format": "ObjectId",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "example": "Usuario Uno"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "uno@example.com"
          },
          "rol": {
            "$ref": "#/components/schemas/RoleEnum"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "pending"
            ],
            "example": "active"
          },
          "profileImage": {
            "type": "string",
            "format": "url",
            "nullable": true
          },
          "delete_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "readOnly": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
      "User": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UserBase"
          }
        ]
      },
      "UserListApiResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/User"
        }
      },
      "UserRegisterPayload": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Nuevo Usuario Tres"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "tres@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "password123"
          },
          "rol": {
            "$ref": "#/components/schemas/RoleEnum"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "rol"
        ]
      },
      "UserRegisterSuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "Solicitud de registro enviada. Pendiente de autorización."
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "LoginPayload": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "uno@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "password_de_uno"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserLoginSuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "Usuario logueado correctamente"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
      "UserUpdatePayload": {
        "type": "object",
        "description": "Campos actualizables para un usuario. Todos son opcionales.",
        "properties": {
          "name": {
            "type": "string"
          },
          "rol": {
            "$ref": "#/components/schemas/RoleEnum"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "pending"
            ]
          }
        }
      },
      "UserUpdateSuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "Usuario actualizado correctamente"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "UserDeleteSuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "Usuario marcado como eliminado"
          },
          "user": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "rol": {
                "$ref": "#/components/schemas/RoleEnum"
              },
              "delete_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
      "RequestPasswordResetPayload": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "uno@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "RequestPasswordResetResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Si tu correo está registrado y activo, recibirás un email con instrucciones."
          }
        }
      },
      "ResetPasswordPayload": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "uno@example.com"
          },
          "newPassword": {
            "type": "string",
            "format": "password"
          },
          "confirmPassword": {
            "type": "string",
            "format": "password"
          }
        },
        "required": [
          "email",
          "newPassword",
          "confirmPassword"
        ]
      },
      "ResetPasswordResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Contraseña actualizada con éxito."
          }
        }
      },
      "ProfileImageUploadResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Imagen de perfil actualizada."
          },
          "imageUrl": {
            "type": "string",
            "format": "url",
            "example": "/uploads/profiles/usuario@example.com-timestamp.jpg"
          }
        }
      }, 
...
    },
    "responses": {
      "InternalServerError": {
        "description": "Error interno del servidor.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "serverError": {
                "value": {
                  "status": "error",
                  "message": "Ocurrió un error inesperado en el servidor."
                }
              }
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "No autorizado. Token faltante, inválido o expirado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "noToken": {
                "value": {
                  "status": "fail",
                  "message": "No se proporcionó token de acceso."
                }
              },
              "invalidToken": {
                "value": {
                  "status": "fail",
                  "message": "Token inválido o expirado."
                }
              }
            }
          }
        }
      },
...
      "NotFoundError": {
        "description": "Recurso no encontrado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "resourceNotFound": {
                "value": {
                  "status": "fail",
                  "message": "El recurso solicitado no fue encontrado."
                }
              }
            }
          }
        }
      },
      "BadRequestError": {
        "description": "Solicitud incorrecta o datos de entrada inválidos.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "validationError": {
                "value": {
                  "status": "fail",
                  "message": "Error de validación.",
                  "details": "El campo 'nombre' no puede estar vacío."
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
    "/user/register": {
      "post": {
        "tags": [
          "User Auth & Management"
        ],
        "summary": "Registrar un nuevo usuario",
        "description": "Crea una nueva cuenta de usuario en el sistema.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterPayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario registrado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRegisterSuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User Auth & Management"
        ],
        "summary": "Iniciar sesión de usuario",
        "description": "Autentica a un usuario con email y contraseña y devuelve un token de acceso.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autenticación exitosa.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserLoginSuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/user/authorize/{token}/{action}": {
      "get": {
        "tags": [
          "User Auth & Management"
        ],
        "summary": "Procesar activación o rechazo de cuenta de usuario",
        "description": "Usado generalmente desde un enlace enviado por email para activar o rechazar una cuenta pendiente.",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "description": "Token de autorización enviado al usuario.",
            "schema": {
              "type": "string",
              "example": "longSecureTokenString12345"
            }
          },
          {
            "name": "action",
            "in": "path",
            "required": true,
            "description": "Acción a realizar.",
            "schema": {
              "type": "string",
              "enum": [
                "approve",
                "reject"
              ],
              "example": "approve"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Acción procesada exitosamente. (Redirige o muestra mensaje HTML).",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "example": "<h1>Tu cuenta ha sido activada.</h1>"
                }
              }
            }
          },
          "400": {
            "description": "Token inválido o expirado, o acción no válida.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "example": "<h1>Error: El enlace no es válido o ha expirado.</h1>"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/user/update/{email}": {
      "put": {
        "tags": [
          "User Auth & Management"
        ],
        "summary": "Actualizar información de un usuario",
        "description": "Permite actualizar campos de un usuario existente, identificado por su email.",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "description": "Email del usuario a actualizar.",
            "schema": {
              "type": "string",
              "format": "email",
              "example": "ana.gomez@example.com"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdatePayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserUpdateSuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/user/delete/{email}": {
      "post": {
        "tags": [
          "User Auth & Management"
        ],
        "summary": "Marcar un usuario como eliminado (soft delete)",
        "description": "Cambia el estado de un usuario para marcarlo como eliminado.",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "description": "Email del usuario a eliminar.",
            "schema": {
              "type": "string",
              "format": "email",
              "example": "usuario.a.eliminar@example.com"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario marcado como eliminado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDeleteSuccessResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/user/auth/editpws": {
      "post": {
        "tags": [
          "User Auth & Management"
        ],
        "summary": "Solicitar reseteo de contraseña",
        "description": "Inicia el flujo para resetear la contraseña de un usuario.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPasswordResetPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solicitud de reseteo procesada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestPasswordResetResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/user/auth/resetpassword": {
      "post": {
        "tags": [
          "User Auth & Management"
        ],
        "summary": "Establecer nueva contraseña después de recuperación",
        "description": "Permite al usuario establecer una nueva contraseña.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contraseña actualizada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/user/profile/{userEmail}/image": {
      "post": {
        "tags": [
          "User Auth & Management"
        ],
        "summary": "Subir imagen de perfil de usuario",
        "description": "Permite subir o reemplazar la imagen de perfil de un usuario.",
        "parameters": [
          {
            "name": "userEmail",
            "in": "path",
            "required": true,
            "description": "Email del usuario.",
            "schema": {
              "type": "string",
              "format": "email",
              "example": "usuario@example.com"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "profileImageField": {
                    "type": "string",
                    "format": "binary",
                    "description": "El archivo de imagen a subir."
                  }
                },
                "required": [
                  "profileImageField"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Imagen subida y guardada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileImageUploadResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/user/prueba": {
      "get": {
        "tags": [
          "User Auth & Management"
        ],
        "summary": "Obtener lista de todos los usuarios (prueba/admin)",
        "description": "Devuelve una lista de todos los usuarios registrados.",
        "responses": {
          "200": {
            "description": "Lista de usuarios obtenida exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListApiResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
...
}